simplification: workers - observer class removed
There are only workers!
We can control if a worker gets the previous thread or starts a new one and whether they have user interaction or not.
This way we can recreate the Observer functionality and can even switch between assistants on the same thread.
It's cleaner.
Added comments to the reworked config file and works nicely now.